requirements.txt
pysocks
requests
pystyle
Pillow
colorama
opencv-python
numpy
